Refactor : SettingActivty / AuthRepository 생성 후 로그아웃 이동
