+++ app/src/main/java/com/nbcproject/earlybirdy/repository/AuthRepository.kt
@@ -1,0 +1,5 @@
+package com.nbcproject.earlybirdy.repository
+
+interface AuthRepository {
+    fun signOut()
+}
