+++ app/src/main/java/com/nbcproject/earlybirdy/setting/SettingActivity.kt
@@ -9,21 +9,25 @@ import android.os.Bundle
 import android.provider.Settings
 import androidx.activity.viewModels
 import androidx.annotation.RequiresApi
+import androidx.lifecycle.ViewModelProvider
 import com.nbcproject.earlybirdy.R
 import com.nbcproject.earlybirdy.databinding.ActivitySettingBinding
 import com.nbcproject.earlybirdy.main.MainActivity
+import com.nbcproject.earlybirdy.my_page.MyPageViewModel
+import com.nbcproject.earlybirdy.my_page.MyPageViewModelFactory
+import com.nbcproject.earlybirdy.setting.dialog.SettingDeleteDialog
+import com.nbcproject.earlybirdy.setting.viewmodel.SettingViewModel
+import com.nbcproject.earlybirdy.setting.viewmodel.SettingViewModelFactory
 import com.nbcproject.earlybirdy.util.Constants.Companion.genralconditionUrl
 import com.nbcproject.earlybirdy.util.Constants.Companion.openlicenseUrl
 import com.nbcproject.earlybirdy.util.navigateToSigninActivity
 import com.nbcproject.earlybirdy.util.showToast
-
 import com.nbcproject.earlybirdy.util.Constants.Companion.supportUrl
 
 class SettingActivity : MainActivity() {
     private lateinit var binding: ActivitySettingBinding
     private lateinit var settingDeleteDialog: SettingDeleteDialog
-    private val settingViewModel: SettingViewModel by viewModels()
-
+    private lateinit var settingViewModel : SettingViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -31,15 +35,16 @@ class SettingActivity : MainActivity() {
         setContentView(binding.root)
 
         settingDeleteDialog = SettingDeleteDialog(this@SettingActivity)
+        settingViewModel = ViewModelProvider(this, SettingViewModelFactory())[SettingViewModel::class.java]
 
         setOnclickListener()
-
-        //화면전환 애니메이션
-        overridePendingTransition(R.anim.slide_right_enter, R.anim.slide_left_exit)
-
     }
 
     private fun setOnclickListener() {
+        //알림 설정 버튼
+        binding.btnNotificationSettings.setOnClickListener {
+            presentNotificationSetting(this)
+        }
 
         //고객지원 구글폼으로 연결
         binding.btnUserSupport.setOnClickListener {
@@ -53,13 +58,7 @@ class SettingActivity : MainActivity() {
 
         //약관 구글폼으로 연결
         binding.btnGenralCondition.setOnClickListener {
-            intentToGenralconditionLink()
-        }
-
-        //알림 설정 버튼
-        binding.btnNotificationSettings.setOnClickListener {
-            presentNotificationSetting(this)
-
+            intentToGeneralConditionLink()
         }
 
         //로그아웃 버튼
@@ -67,56 +66,16 @@ class SettingActivity : MainActivity() {
             signOut()
         }
 
+        //회원 탈퇴 버튼
+        binding.btnDeleteAccount.setOnClickListener {
+            settingDeleteDialog.show()
+        }
+
         //뒤로가기 버튼
         binding.ivBack.setOnClickListener {
             finish()
         }
 
-        //회원 탈퇴 버튼
-        binding.btnDeleteAccount.setOnClickListener {
-            settingDeleteDialog.show()
-        }
-    }
-
-    private fun intentToSupportLink() {
-        val supportLink = supportUrl
-
-        val browserIntent = Intent(Intent.ACTION_VIEW, Uri.parse(supportLink))
-        try {
-            startActivity(browserIntent)
-        } catch (e: ActivityNotFoundException) {
-            showToast(this, "연결 오류 발생")
-        }
-    }
-
-    private fun intentToOpenLicenseUrlLink() {
-        val supportLink = supportUrl
-
-        val browserIntent = Intent(Intent.ACTION_VIEW, Uri.parse(openlicenseUrl))
-        try {
-            startActivity(browserIntent)
-        } catch (e: ActivityNotFoundException) {
-            showToast(this, "연결 오류 발생")
-        }
-    }
-
-    private fun intentToGenralconditionLink() {
-        val supportLink = supportUrl
-
-        val browserIntent = Intent(Intent.ACTION_VIEW, Uri.parse(genralconditionUrl))
-        try {
-            startActivity(browserIntent)
-        } catch (e: ActivityNotFoundException) {
-            showToast(this, "연결 오류 발생")
-        }
-    }
-
-    //로그아웃
-    private fun signOut() {
-        settingViewModel.signOut()
-        showToast(this, "로그아웃 되었습니다. 로그인 페이지로 이동합니다.")
-        finish()
-        navigateToSigninActivity(this)
     }
 
     //Notification Setting Page 이동 함수
@@ -131,6 +90,41 @@ class SettingActivity : MainActivity() {
         } catch (e: ActivityNotFoundException) {
             e.printStackTrace()
         }
+    }
+
+    private fun intentToSupportLink() {
+        val supportLink = supportUrl
+
+        val browserIntent = Intent(Intent.ACTION_VIEW, Uri.parse(supportLink))
+        try {
+            startActivity(browserIntent)
+        } catch (e: ActivityNotFoundException) {
+            showToast(this, R.string.setting_connect_error_toast.toString())
+        }
+    }
+
+    private fun intentToOpenLicenseUrlLink() {
+        val browserIntent = Intent(Intent.ACTION_VIEW, Uri.parse(openlicenseUrl))
+        try {
+            startActivity(browserIntent)
+        } catch (e: ActivityNotFoundException) {
+            showToast(this, R.string.setting_connect_error_toast.toString())
+        }
+    }
+
+    private fun intentToGeneralConditionLink() {
+        val browserIntent = Intent(Intent.ACTION_VIEW, Uri.parse(genralconditionUrl))
+        try {
+            startActivity(browserIntent)
+        } catch (e: ActivityNotFoundException) {
+            showToast(this, R.string.setting_connect_error_toast.toString())
+        }
+    }
+
+    //로그아웃
+    private fun signOut() {
+        settingViewModel.signOut()
+        navigateToSigninActivity(this)
     }
 
     //오레오 버전 이상부터 사용 가능
