+++ app/src/main/java/com/nbcproject/earlybirdy/setting/viewmodel/SettingViewModel.kt
@@ -1,0 +1,11 @@
+package com.nbcproject.earlybirdy.setting.viewmodel
+
+import android.util.Log
+import androidx.lifecycle.ViewModel
+import com.nbcproject.earlybirdy.repository.AuthRepositoryImpl
+
+class SettingViewModel(private val authRepository: AuthRepositoryImpl) : ViewModel() {
+    fun signOut() {
+        authRepository.signOut()
+    }
+}
