+++ app/src/main/java/com/nbcproject/earlybirdy/setting/viewmodel/SettingViewModelFactory.kt
@@ -1,0 +1,15 @@
+package com.nbcproject.earlybirdy.setting.viewmodel
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import com.nbcproject.earlybirdy.repository.AuthRepositoryImpl
+
+class SettingViewModelFactory : ViewModelProvider.Factory {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(SettingViewModel::class.java)) {
+            val authRepository = AuthRepositoryImpl()
+            return SettingViewModel(authRepository) as T
+        }
+        throw IllegalArgumentException("Unknown ViewModel class")
+    }
+}
