+++ app/src/main/res/values/strings.xml
@@ -220,7 +220,11 @@
     <string name="itemranking_nickname">닉네임</string>
     <string name="itemranking_exp">경험치량</string>
     
-    <!--Board 갤러리 접근 권한 관련-->
+    <!--Setting Activity Toast 메세지-->
+    <string name="setting_connect_error_toast">"연결 오류 발생</string>
+    <string name="setting_logout_toast">로그아웃 되었습니다. 로그인 페이지로 이동합니다.</string>
+
+    <!--Board 권한 설정-->
     <string name="modify_permissionRequired"></string>
     <string name="modify_profilePermission"></string>
     <string name="modify_agree"></string>
